demand-forecasting/requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
shap>=0.42.0
fastapi>=0.95.0
uvicorn>=0.22.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.0
pydantic>=1.10.0
requests>=2.31.0
joblib>=1.2.0
psutil>=5.9.0
